Testing Flake8-Check With A Bad Syntaxed Python file
A new rule to the `master` branch has been added
to Github, such that Flake8's checks must pass
before merging to the `master`.

We have now created the file: `bad.py`,
that has a bad python syntax, to check if the
new rule on the `master` really protects it.